requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
protobuf
sentencepiece
sacremoses
python-jose[cryptography]